third_party/WebKit/LayoutTests/custom-properties/registered-property-computation-expected.txt

This is a testharness.js-based test.
FAIL <length> values are computed correctly for divWithFontSizeSet assert_equals: expected "140px" but got "224px"
FAIL <length-percentage> values are computed correctly for divWithFontSizeSet assert_equals: expected "170px" but got "272px"
FAIL <length>+ values are computed correctly for divWithFontSizeSet assert_equals: expected "10px 30px" but got "10px 48px"
FAIL <length-percentage>+ values are computed correctly for divWithFontSizeSet assert_equals: expected "calc(10px + 50%) 4px" but got "calc(16px + 50%) 4px"
PASS <length> values are computed correctly for divWithFontSizeInherited 
PASS <length-percentage> values are computed correctly for divWithFontSizeInherited 
PASS <length>+ values are computed correctly for divWithFontSizeInherited 
PASS <length-percentage>+ values are computed correctly for divWithFontSizeInherited 
Harness: the test ran to completion.
